Namespaces API files that might change in future under V1 * common files (i.e. middleware) remain under /api - not namespaced w/ version
