--- spec/flipper/api_spec.rb
@@ -1,6 +1,0 @@
-require 'helper'
-
-RSpec.describe Flipper::Api do
-  describe 'Initializing middleware with flipper instance' do
-  end
-end
